refactor(car): car 도메인 관련 함수 분리
